+++ src/widgets/nextpage-smooth-scroll/index.ts
@@ -1,1 +1,0 @@
-export { NextPageSmoothScroll } from './ui/nextpage-smooth-scroll'